Separate initialize, update and destroy functions.
